requirements.txt
google.generativeai
flask
streamlit==1.16.0
PyPDF2==1.26.0
scikit-learn==1.1.3
requests==2.28.1
beautifulsoup4==4.11.1
python-dotenv==0.21.0
spacy==3.5.0